+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.1.0
 
+#### 💥 Breaking
+
+- Will now longer check `engines` in `package.json` when detecting a version.
+
 #### 🚀 Updates
 
 - Updated to support proto v0.14 release.